Conflicts:
	HuntingBrave_client/IEX2010.v12.suo

+++ HuntingBrave_client/source/GameManager.h
@@ -26,7 +26,7 @@ class GameManager : public Singleton<GameManager>
 	friend Singleton<GameManager>;
 
 public:
-	int		timer;
+	float		timer;
 	bool	isComplete;		//Ú‘±‚É€”õŠ®—¹‚µ‚½‚©‚Ç‚¤‚©
 
 
@@ -45,10 +45,10 @@ public:
 	void	Update( void );
 
 	//	î•ñİ’è
-	void	SetTimer( int time ){ timer = time; }
+	void	SetTimer( float time ){ timer = time; }
 
 	//	î•ñæ“¾
-	int		GetTime( void ){ return timer; }
+	float		GetTime( void ){ return timer; }
 	bool		GetIsComplete(void){ return isComplete; }
 };
